Add auto healing and auto scaling Managed Instance Group
This patch establishes a generic, reusable module that takes in a
startup script.  A regional, highly available managed instance group is
created with auto-healing and auto-scaling health checks.

An internal load balancer IP is allocated and assigned.

The startup script can be anything, but the initial use case is to set
up an HTTP forward proxy.