Add GCP Ops Agent
Reconfigure Apache2 to prevent log spam to the /healthz endpoint and
also configure the Ops Agent to parse the slightly different variation
between the main access log and the vhost access log.  The notable
difference is the first field is the host for the vhost access logs.

The default Ops Agent pipeline is turned off to reduce log spam,
particularly the health checks to the hc-traffic service.